+++ zcash_client_backend/src/data_api.rs
@@ -4923,6 +4923,8 @@ mod tests {
             Some(orchard).flatten(),
             Some(sapling).flatten(),
             transparent,
+            None,
+            None,
         )
         .expect("test UA must be valid")
         .into()
@@ -5134,6 +5136,8 @@ mod tests {
         let frankenstein = UnifiedAddress::from_receivers(
             Some(ua2.orchard().copied().expect("orchard receiver")),
             Some(ua1.sapling().copied().expect("sapling receiver")),
+            None,
+            None,
             None,
         )
         .expect("sapling+orchard UA must be valid");