+++ zcash_client_sqlite/src/wallet/init/migrations/ufvk_support.rs
@@ -159,7 +159,10 @@ impl<P: consensus::Parameters> RusqliteMigration for Migration<P> {
                 seed_is_relevant = true;
 
                 let ufvk_str: String = ufvk.encode(&self.params);
-                let address_str: String = ufvk.default_address(ua_request)?.0.encode(&self.params);
+                let address_str: String = ufvk
+                    .default_address(ua_request)?
+                    .0
+                    .encode_receiver_preserving(&self.params);
 
                 // This migration, and the wallet behaviour before it, stored the default
                 // transparent address in the `accounts` table. This does not necessarily